Preliminary user management
This commit lays groundwork on how persistent objects will be managed by the server.
The index.php page creates a new user, saves it in the database, checks the password twice (once with an incorrect, and then with a correct password), and then attempts to do the same with the same user constructed using saved data from database.
Passwords are automatically hashed so nobody can see the password even if the database is accessed.