Login system
Added an endpoint system.
Endpoints return dffierent response codes, which are sent by JavaScript (JQuery)'s AJAX support and are received and handled by displaying appropriate messages or redirecting to another page.
Added register, login and logout pages. Successfully registering redirects the user to the login page, and successfully logging in redirects the user to the main page.
All passwords must have at least 6 characters.
The PHP session system is used for storing the information about who are logged in as. This information is displayed on the main page.
The test page has been extracted to a separate page instead of performing tests on the main page.